feat: separated drag-t-select logic to hook
