+++ app/components/ExplorerWindow.jsx
@@ -1,225 +1,81 @@
 "use client";
-import { useRef, useState, useEffect } from "react";
+import { useEffect, useRef, useState } from "react";
+import MenuBar from "./catalog/MenuBar";
+import AddressBar from "./catalog/AddressBar";
+import WindowContainer from "./core/WindowContainer";
+import TitleBar from "./core/TitleBar";
+import FilterBar from "./catalog/FilterBar";
+import IncidentCard from "./catalog/IncidentCard";
+import YearFolder from "./catalog/YearFolder";
+import SelectionBox from "./catalog/SelectionBox";
+import useIncidentProcessor from "../hooks/useIncidentProcessor";
+import useIncidentFilter from "../hooks/useIncidentFilter";
+import useViewManager from "../hooks/useViewManager";
+import useSelectionManager from "../hooks/useSelectionManager";
 
-export default function ExplorerWindow({
+const ExplorerWindow2 = ({
   incidents,
   selectedIncidents,
   setSelectedIncidents,
   setDisplayedIncident,
   setContextMenu,
-}) {
-  const [lastClickedIndex, setLastClickedIndex] = useState(null);
-  const containerRef = useRef(null);
-  const [isSelecting, setIsSelecting] = useState(false);
-  const [startPoint, setStartPoint] = useState({ x: 0, y: 0 });
-  const [selectionBox, setSelectionBox] = useState({
-    x: 0,
-    y: 0,
-    width: 0,
-    height: 0,
-  });
+}) => {
+  const { incidentsByDecade, decades } = useIncidentProcessor(incidents);
 
-  const [currentView, setCurrentView] = useState("years");
-  const [currentYear, setCurrentYear] = useState(null);
+  const {
+    activeFilter,
+    searchQuery,
+    filteredIncidents,
+    handleFilterClick,
+    handleSearchChange,
+    handleSearchClear,
+  } = useIncidentFilter(incidents);
 
-  const incidentsByYear = incidents.reduce((acc, incident) => {
-    const year = new Date(incident.incident_date).getFullYear();
-    if (!acc[year]) {
-      acc[year] = [];
-    }
-    acc[year].push(incident);
-    return acc;
-  }, {});
+  const {
+    currentView,
+    currentYear,
+    setCurrentView,
+    setCurrentYear,
+    filteredDecades,
+    visibleIncidents,
+    navigateToYear,
+    navigateToRoot,
+    currentPath,
+    windowTitle,
+  } = useViewManager(incidents, incidentsByDecade, filteredIncidents);
 
-  const years = Object.keys(incidentsByYear).sort();
-
-  const handleItemClick = (e, incident, index) => {
-    e.stopPropagation();
-    if (e.shiftKey && lastClickedIndex !== null) {
-      const start = Math.min(lastClickedIndex, index);
-      const end = Math.max(lastClickedIndex, index);
-      const range = incidents.slice(start, end + 1);
-      setSelectedIncidents(range);
-    } else if (e.ctrlKey || e.metaKey) {
-      if (selectedIncidents.includes(incident)) {
-        setSelectedIncidents(selectedIncidents.filter((i) => i !== incident));
-      } else {
-        setSelectedIncidents([...selectedIncidents, incident]);
-      }
-      setLastClickedIndex(index);
-    } else {
-      setSelectedIncidents([incident]);
-      setLastClickedIndex(index);
-    }
-  };
+  const {
+    containerRef,
+    isSelecting,
+    selectionBox,
+    handleItemClick,
+    handleItemContextMenu,
+    handleMouseDown,
+    handleMouseMove,
+    handleMouseUp,
+  } = useSelectionManager(setSelectedIncidents);
 
   const handleItemDoubleClick = (incident) => {
     setDisplayedIncident(incident);
   };
 
-  const handleFolderDoubleClick = (year) => {
-    setCurrentView("incidents");
-    setCurrentYear(year);
+  const handleFolderDoubleClick = (decade) => {
+    navigateToYear(decade);
     setSelectedIncidents([]);
   };
 
   const handleBackClick = () => {
-    setCurrentView("years");
-    setCurrentYear(null);
+    navigateToRoot();
     setSelectedIncidents([]);
-  };
-
-  const handleItemContextMenu = (e, incident, index) => {
-    e.preventDefault();
-    e.stopPropagation();
-
-    if (!selectedIncidents.includes(incident)) {
-      setSelectedIncidents([incident]);
-      setLastClickedIndex(index);
-    }
-
-    setContextMenu({
-      visible: true,
-      x: e.clientX,
-      y: e.clientY,
-      onFile: true,
-      incidents: selectedIncidents.includes(incident)
-        ? selectedIncidents
-        : [incident],
-    });
-  };
-
-  const handleMouseDown = (e) => {
-    if (e.button !== 0 || e.shiftKey || e.ctrlKey || e.metaKey) return;
-
-    setIsSelecting(true);
-    const rect = containerRef.current.getBoundingClientRect();
-    const x = e.clientX - rect.left;
-    const y = e.clientY - rect.top;
-    setStartPoint({ x, y });
-    setSelectionBox({ x, y, width: 0, height: 0 });
-
-    setSelectedIncidents([]);
-  };
-
-  const handleMouseMove = (e) => {
-    if (!isSelecting) return;
-    e.preventDefault();
-
-    const rect = containerRef.current.getBoundingClientRect();
-    const currentX = e.clientX - rect.left;
-    const currentY = e.clientY - rect.top;
-
-    const newBox = {
-      x: Math.min(startPoint.x, currentX),
-      y: Math.min(startPoint.y, currentY),
-      width: Math.abs(currentX - startPoint.x),
-      height: Math.abs(currentY - startPoint.y),
-    };
-    setSelectionBox(newBox);
-
-    const newlySelected = [];
-    const items =
-      currentView === "years" ? years : incidentsByYear[currentYear];
-
-    for (let i = 0; i < items.length; i++) {
-      const itemEl = document.getElementById(
-        `${currentView === "years" ? "folder" : "incident"}-${i}`
-      );
-      if (!itemEl) continue;
-      const itemRect = itemEl.getBoundingClientRect();
-
-      const itemBox = {
-        x: itemRect.left - rect.left,
-        y: itemRect.top - rect.top,
-        width: itemRect.width,
-        height: itemRect.height,
-      };
-
-      if (boxesIntersect(newBox, itemBox)) {
-        if (currentView === "years") {
-          newlySelected.push(years[i]);
-        } else {
-          newlySelected.push(incidentsByYear[currentYear][i]);
-        }
-      }
-    }
-    setSelectedIncidents(newlySelected);
-  };
-
-  const handleMouseUp = (e) => {
-    if (e.button !== 0) return;
-    setIsSelecting(false);
-  };
-
-  function boxesIntersect(a, b) {
-    return !(
-      a.x + a.width < b.x ||
-      a.x > b.x + b.width ||
-      a.y + a.height < b.y ||
-      a.y > b.y + b.height
-    );
-  }
-
-  const [activeFilter, setActiveFilter] = useState(null);
-  const [searchQuery, setSearchQuery] = useState("");
-
-  const handleFilterClick = (category) => {
-    if (activeFilter === category) {
-      setActiveFilter(null);
-    } else {
-      setActiveFilter(category);
-    }
-  };
-
-  const handleSearchChange = (e) => {
-    setSearchQuery(e.target.value);
   };
 
   const handleSearchSubmit = (e) => {
     e.preventDefault();
   };
 
-  const filteredIncidents = incidents.filter((incident) => {
-    if (activeFilter && incident.category !== activeFilter) {
-      return false;
-    }
-
-    if (searchQuery.trim()) {
-      const query = searchQuery.toLowerCase();
-      return (
-        incident.name.toLowerCase().includes(query) ||
-        incident.category.toLowerCase().includes(query) ||
-        (incident.severity && incident.severity.toLowerCase().includes(query))
-      );
-    }
-
-    return true;
-  });
-
-  const filteredYears =
-    searchQuery || activeFilter
-      ? years.filter((year) => {
-          return incidentsByYear[year].some((incident) =>
-            filteredIncidents.includes(incident)
-          );
-        })
-      : years;
-
-  const visibleIncidents = currentYear
-    ? incidentsByYear[currentYear].filter((incident) =>
-        filteredIncidents.includes(incident)
-      )
-    : [];
-
   return (
-    <div
+    <WindowContainer
       ref={containerRef}
-      className="explorer-container h-[80vh] w-[75vw] shadow-win95"
-      style={{
-        userSelect: "none",
-        position: "relative",
-      }}
       onMouseDown={handleMouseDown}
       onMouseMove={handleMouseMove}
       onMouseUp={handleMouseUp}
@@ -234,107 +90,61 @@ export default function ExplorerWindow({
         });
       }}
     >
-      <div className="explorer-window-bar">
-        <div className="folder-name">
-          <img src="/win95-folder-icon.png" alt="Folder Icon" />
-          <p>Technology Incidents{currentYear ? ` - ${currentYear}` : ""}</p>
-        </div>
-        <div className="window-buttons">
-          <div id="min-button"></div>
-          <div id="max-button"></div>
-          <div id="close-button"></div>
-        </div>
-      </div>
+      <TitleBar
+        title={windowTitle}
+        icon="/win95-folder-icon.png"
+        onMinimize={() => {
+          /* handle minimize */
+        }}
+        onMaximize={() => {
+          /* handle maximize */
+        }}
+        onClose={() => {
+          /* handle close */
+        }}
+      />
+      <MenuBar />
+      <AddressBar
+        currentPath={currentPath}
+        searchQuery={searchQuery}
+        onSearchChange={handleSearchChange}
+        onSearchSubmit={(e) => e.key === "Enter" && handleSearchSubmit(e)}
+      />
 
-      <div className="explorer-header">
-        <div className="menu">
-          <p>File</p>
-          <p>Edit</p>
-          <p>View</p>
-          <p>Help</p>
-        </div>
-        <div className="path-display">
-          <p>Address</p>
-          <div className="explorer-path flex-1 flex items-center">
-            <p className="whitespace-nowrap">
-              C:\Technology Incidents{currentYear ? `\\${currentYear}` : ""}\
-            </p>
-            <input
-              type="text"
-              className="flex-1 outline-none border-none bg-transparent"
-              placeholder="Search..."
-              value={searchQuery}
-              onChange={handleSearchChange}
-              onKeyDown={(e) => e.key === "Enter" && handleSearchSubmit(e)}
-            />
-          </div>
-        </div>
-      </div>
-      <div className="filter-bar">
-        <div className="flex flex-row gap-3 items-center justify-start">
-          {currentView === "incidents" && (
-            <button
-              className="win95-folder-nav-button"
-              onClick={handleBackClick}
-            >
-              <div className="arrow-left"></div>
-            </button>
-          )}
-          <p>Category:</p>
-          {[...new Set(incidents.map((incident) => incident.category))].map(
-            (category) => (
-              <button
-                key={category}
-                className={`filter-button ${activeFilter === category ? "bg-win95blue text-white" : "hover:bg-win95blue hover:text-white"}`}
-                onClick={() => handleFilterClick(category)}
-              >
-                {category}
-              </button>
-            )
-          )}
-        </div>
-      </div>
+      <FilterBar
+        currentView={currentView}
+        categories={[
+          ...new Set(incidents.map((incident) => incident.category)),
+        ]}
+        activeFilter={activeFilter}
+        onFilterClick={handleFilterClick}
+        onBackClick={handleBackClick}
+      />
 
       <div className="explorer-content">
         {currentView === "years" ? (
-          filteredYears.map((year, index) => (
-            <div
-              key={year}
-              id={`folder-${index}`}
-              className="flex flex-col items-center justify-center gap-1 p-2 cursor-pointer hover:bg-blue-100"
-              onDoubleClick={() => handleFolderDoubleClick(year)}
-            >
-              <img src="/win95-folder-icon.png" className="w-8 h-8" />
-              <p>{year}</p>
-              <p className="text-xs text-gray-500">
-                {incidentsByYear[year].length} incidents
-              </p>
-            </div>
+          filteredDecades.map((decade, index) => (
+            <YearFolder
+              key={decade}
+              year={`${decade}s`}
+              incidentCount={incidentsByDecade[decade]?.length || 0}
+              index={index}
+              onDoubleClick={() => handleFolderDoubleClick(decade)}
+            />
           ))
         ) : visibleIncidents.length > 0 ? (
           visibleIncidents.map((entry, index) => {
             const isSelected = selectedIncidents.includes(entry);
             return (
-              <div
-                key={entry.id}
-                id={`incident-${index}`}
-                className={`case-container ${isSelected ? "selected" : ""}`}
-                style={{ position: "relative" }}
+              <IncidentCard
+                key={entry.id || index}
+                entry={entry}
+                index={index}
+                isSelected={isSelected}
                 onClick={(e) => handleItemClick(e, entry, index)}
                 onDoubleClick={() => handleItemDoubleClick(entry)}
                 onContextMenu={(e) => handleItemContextMenu(e, entry, index)}
-              >
-                <div className="casefile">
-                  <img src="/computer.png" className="w-12 h-12" />
-                  <div
-                    className={`absolute bottom-0 right-0 w-3 h-3 rounded-full ${entry.severity.toLowerCase()}`}
-                  />
-                </div>
-                <p>{entry.name}</p>
-                <p className="text-xs bg-gray-300 px-2 category-label">
-                  {entry.category}
-                </p>
-              </div>
+              />
             );
           })
         ) : (
@@ -344,20 +154,7 @@ export default function ExplorerWindow({
         )}
       </div>
 
-      {isSelecting && (
-        <div
-          style={{
-            position: "absolute",
-            border: "1px dashed black",
-            background: "transparent",
-            left: selectionBox.x,
-            top: selectionBox.y,
-            width: selectionBox.width,
-            height: selectionBox.height,
-            pointerEvents: "none",
-          }}
-        />
-      )}
-    </div>
+      <SelectionBox isSelecting={isSelecting} selectionBox={selectionBox} />
+    </WindowContainer>
   );
 }
